# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Advancia Smart Wallets API
  version: "1.0.0"
  description: |
    REST API for the Advancia Smart Wallets platform — a fintech super-app
    with embedded smart wallets (ERC-4337), payments, healthcare, banking,
    loyalty, subscriptions, and admin management.

    ## Authentication
    Most endpoints require a JWT session token obtained from `/api/auth/*`.
    Pass it as a cookie (`session`) or `Authorization: Bearer <token>` header.

    Admin endpoints require a separate admin session via `/api/admin/login`.

    ## Versioning
    All endpoints are available under both `/api/*` and `/api/v1/*`.
    The `X-API-Version` response header indicates the active version.

    ## Rate Limiting
    Auth endpoints (`/api/auth/*`) are rate-limited to 20 requests/minute per IP.
    Exceeding the limit returns `429 Too Many Requests` with a `Retry-After` header.
  contact:
    name: Advancia Dev Team
    email: security@advancia.health
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://advancia-healthcare1.vercel.app
    description: Production (Vercel)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
    cronAuth:
      type: http
      scheme: bearer
      description: CRON_SECRET bearer token for cron endpoints
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Invalid request body"
...
  responses:
...
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
security:
  - bearerAuth: []
  - cookieAuth: []

# ═══════════════════════════════════════════════════════════════
#  PATHS — grouped by tag / domain
# ═══════════════════════════════════════════════════════════════
paths:
...
  /api/auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email with token
      security: []
      responses:
        "200": { description: "Email verified" }
        "400": { $ref: "#/components/responses/BadRequest" }

  # ─── Wallets ───────────────────────────────────
...
  /api/gift-cards:
    get:
      tags: [Gift Cards]
      summary: List gift cards (paginated)
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [ACTIVE, REDEEMED, EXPIRED] }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200": { description: "Paginated gift cards" }
    post:
      tags: [Gift Cards]
      summary: Purchase a gift card
      responses:
        "201": { description: "Gift card purchased" }
    patch:
      tags: [Gift Cards]
      summary: Redeem a gift card
      responses:
        "200": { description: "Gift card redeemed" }
...
  /api/cron/notifications:
    get:
      tags: [Cron]
      summary: Process queued push notifications
      security: [{ cronAuth: [] }]
      responses:
        "200": { description: "Sent count" }
...